TPAMI/requirements.txt
torch>=1.7.0
torchvision>=0.8.0
numpy>=1.19.0
pillow>=8.0.0
rawpy>=0.16.0
einops>=0.3.0
